Get all the articles and Fix up the style of doc

+++ tools/MyHTMLParser.py
@@ -44,42 +44,53 @@ class XQHTMLParser(HTMLParser):
         self.myclient = HTMLClient()
         self.text = ''
         self.title = False
+        self.isdes = False
         self.picList=[]
     def handle_starttag(self, tag, attrs):
         #print "Encountered the beginning of a %s tag" % tag
+        self.title = False
+        self.isdes = False
         if re.match(r'h(\d)', tag):
             self.title = True 
+            print(tag)
         if tag == "img":
             if len(attrs) == 0: pass
             else:
                 for (variable, value)  in attrs:
                     if variable == "src":
-                        picdata = self.myclient.GetPic(value)
-                        pictmp = value.split('/')[-1]
-                        with open(pictmp, 'wb') as pic:
-                            pic.write(bytes(picdata))
-                            pic.close()
-                        self.doc.add_picture(pictmp, width=Inches(1.25))
-                        self.picList.append(pictmp)
+                        picdata = self.myclient.GetPic(value.split('!')[0])
+                        pictmp = value.split('/')[-1].split('!')[0]
+                        if pictmp.find('jpg') < 0:
+                            with open(pictmp, 'wb') as pic:
+                                pic.write(bytes(picdata))
+                                pic.close()
+                            self.doc.add_picture(pictmp, width=Inches(2.25))
+                            self.picList.append(pictmp)
+        if tag == 'script':
+            self.isdes = True
     def handle_data(self, data):
         if self.title == True:
             self.doc.add_paragraph(self.text)
             self.text = ''
-            self.doc.add_heading(data)
-        else:
+            self.doc.add_heading(data, level=2)
+        if self.isdes == False:
             self.text += data
     def handle_endtag(self, tag):
-        if tag == 'br':
-            self.doc.add_paragraph(self.text)
-            self.text = ''
+        #if tag == 'br' or tag == 'p' or tag == 'div':
+        self.doc.add_paragraph(self.text)
+        self.text = ''
     def complete(self, html):
         self.feed(html)
         self.doc.save(self.docfile)
         for item in self.picList:
-            os.remove(item)
+            if os.path.exists(item):
+                os.remove(item)
 
-#xq_parser = XQHTMLParser()
-#xq_parser.feed('<img src="http://xavatar.imedao.com/community/201411/1418865775807-1418865776034.jpeg!50x50.png">')
+#xq_parser = XQHTMLParser("tmp.doc")
+#xq_parser.feed(
+#json = '<img src="http://xqimg.imedao.com/14b069587ccc53fc199a1ca2.jpg"  />'
+#json = '<img src="http://xqimg.imedao.com/14b069587ccc53fc199a1ca2.jpg!custom.jpg" >'
+#json = '<img src="http://xqimg.imedao.com/14b0696fb404e53fd90be7d0.jpg!custom.jpg" >'
 #json = '<img src="http://xavatar.imedao.com/community/201411/1418865775807-1418865776034.jpeg!50x50.png">'
-#xq_parser.complete(json, 'tmp.doc')
+#xq_parser.complete(json)
 #doc.save('tmp.doc')
